requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
opencv-python-headless
Pillow
ultralytics
insightface
scrfd
mediapipe
streamlit
requests
